+++ Crossword.Entities/Crossword.Entities.csproj
@@ -4,7 +4,6 @@
         <TargetFramework>net8.0</TargetFramework>
         <ImplicitUsings>enable</ImplicitUsings>
         <Nullable>enable</Nullable>
-        <RootNamespace>Crosssword.Entities</RootNamespace>
     </PropertyGroup>
 
 </Project>
