+++ Crossword.Entities/CrosswordData.cs
@@ -1,15 +1,4 @@
-﻿//////////////////////////////////////////////////////////////////////////////
-//                                                                          //
-//      Module:     CrosswordData.cs                                           //
-//      Authors:    Aaron Saikovski & Bryan Richards                        //
-//      Date:       23/01/97                                                //
-//      Version:    1.0                                                     //
-//      Purpose:    Utilizes a String Tokenizer to parse the crossword      //
-//                  puzzle components from a data set string.               //
-//                                                                          //
-//////////////////////////////////////////////////////////////////////////////
-
-
+﻿
 namespace Crossword.Entities;
 
 public sealed class CrosswordData
@@ -20,8 +9,11 @@ public sealed class CrosswordData
     public string? PuzzleType { get; set; }
     public int NumCols { get; set; }
     public int NumRows { get; set; }
-    public int NumAcross { get; set; }
-    public int NumDown { get; set; }
+    // public int NumAcross { get; set; }
+    // public int NumDown { get; set; }
+    
+    public int NumAcross { get; }
+    public int NumDown { get;  }
     public int PuzzleId { get; set; }
     public int[]? ColRef { get; set; }
     public int[]? RowRef { get; set; }
@@ -30,13 +22,17 @@ public sealed class CrosswordData
     public string[]? Clues { get; set; }
     public string[]? Answers { get; set; }
     
-    public int[]? Costs { get; set; }
+    //public int[]? Costs { get; set; }
+    public int[]? Costs { get; }
 
     public string? GetLetters { get; set; }
     public string? Blurb { get; set; }
+   
     public int NumQuestions { get; set; }
 
     public int NumBytes { get; set; }
+
+    
 
     #endregion
