+++ Crossword.Parser/GetClues.cs
@@ -1,4 +1,4 @@
-using System.Collections.Generic;
+
 
 namespace Crossword.Parser;
 
@@ -11,7 +11,8 @@ public sealed partial class CrosswordParser
     private void GetClues(IReadOnlyList<string> strData)
     {
         var puzzletempstr = strData[4];
-        string[] cluetemp = puzzletempstr.Split("#");
+        //string[] cluetemp = puzzletempstr.Split("#");
+        var cluetemp = puzzletempstr.Split("#");
         if (_crosswordData == null) return;
         _crosswordData.Clues = new string[_crosswordData.NumQuestions];
         for (var j = 0; j < _crosswordData.NumQuestions; j++)
