+++ Crossword.Parser/GetGridPositions.cs
@@ -1,6 +1,4 @@
 
-
-using System.Collections.Generic;
 using Crossword.Shared.ParserUtils;
 
 namespace Crossword.Parser;
