+++ Crossword.Parser/GetHintLetters.cs
@@ -1,4 +1,4 @@
-using System.Collections.Generic;
+
 
 namespace Crossword.Parser;
