+++ Crossword.Puzzle/Squares/SetLetter.cs
@@ -11,7 +11,14 @@ public sealed partial class Square
     /// </summary>
     /// <param name="letter"></param>
     /// <param name="isAcross"></param>
-    public void SetLetter(char letter, bool isAcross)
+    // public void SetLetter(char letter, bool isAcross)
+    // {
+    //     Letter = letter;
+    //     IsDirty = true;
+    //     ForeColour = UIConstants.SquareHighlightDefault;
+    // }
+    
+    public void SetLetter(char letter)
     {
         Letter = letter;
         IsDirty = true;
