+++ Crossword.Puzzle/Squares/SetObjectRef.cs
@@ -1,4 +1,4 @@
-using System;
+
 using Crossword.Puzzle.ClueAnswerMap;
 using Crossword.Shared.Constants;
