+++ Crossword/EventHandlers/KeyDown.cs
@@ -120,7 +120,8 @@ public sealed partial class CrosswordApp
             //Delete present square's contents if Delete key is pressed
             if (keyInFocus == Keys.Delete)
             {
-                SqCurrentSquare?.SetLetter(' ', IsAcross);
+                //SqCurrentSquare?.SetLetter(' ', IsAcross);
+                SqCurrentSquare?.SetLetter(' ');
             }
         }
         catch (Exception ex)
@@ -145,7 +146,8 @@ public sealed partial class CrosswordApp
             
             //Check to see if a backspace was entered
             if (keyInFocus != Keys.Back) return;
-            SqCurrentSquare?.SetLetter(' ', IsAcross);
+            //SqCurrentSquare?.SetLetter(' ', IsAcross);
+            SqCurrentSquare?.SetLetter(' ');
             SqCurrentSquare = SqCurrentSquare?.GetPrevSq(IsAcross);
             SqCurrentSquare?.GetClueAnswerRef(IsAcross)?.HighlightSquares(SqCurrentSquare, true);
         }
@@ -173,7 +175,8 @@ public sealed partial class CrosswordApp
             if (keyInFocus is < Keys.A or > Keys.Z) return;
             
             //Sets the letter in the current square
-            SqCurrentSquare?.SetLetter(char.ToUpper((char)keyInFocus), IsAcross);
+            //SqCurrentSquare?.SetLetter(char.ToUpper((char)keyInFocus), IsAcross);
+            SqCurrentSquare?.SetLetter(char.ToUpper((char)keyInFocus));
 
             //get next sq or myself(same sq)  if not available
             SqCurrentSquare = SqCurrentSquare?.GetNextSq(IsAcross);
